remove unneeded admin page components from backend
