+++ backend/controller/homepage.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"encoding/json"
-	"html/template"
 	"net/http"
 	"strings"
 
@@ -11,64 +10,7 @@ import (
 	"github.com/julienschmidt/httprouter"
 
 	"github.com/metaclips/LetsTalk/backend/model"
-	"github.com/metaclips/LetsTalk/backend/values"
 )
-
-var homepageTmpl, loginTmpl *template.Template
-
-func init() {
-	var terr error
-	// Use (%%) instead of {{}} for templates.
-	homepageTmpl, terr = template.New("home.html").Delims("(%", "%)").ParseFiles(
-		"backend/views/homepage/home.html",
-		"backend/views/homepage/components/SideBar.vue", "backend/views/homepage/components/ChattingComponent.vue", "backend/views/homepage/components/CallUI.vue")
-
-	if terr != nil {
-		log.Fatalln("error parsing homepage templates", terr)
-	}
-
-	loginTmpl, terr = template.New("login.html").Delims("(%", "%)").ParseFiles("backend/views/loginpage/login.html")
-	if terr != nil {
-		log.Fatalln("error parsing login templates", terr)
-	}
-}
-
-// HomePage is a GET request that validates user credentials and load homepage templates.
-func HomePage(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	data := validatUser(w, r)
-
-	switch data.(type) {
-	case error:
-		log.Println("could not log user in", data)
-		http.Redirect(w, r, "/login", 302)
-
-	default:
-		if err := homepageTmpl.Execute(w, data); err != nil {
-			log.Println(err)
-		}
-	}
-
-}
-
-// HomePageLoginGet loads login page for users to login. Cookies are initially validated if user is already logged in.
-func HomePageLoginGet(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	data := validatUser(w, r)
-
-	// On validate user, if users has initially logged in execute homepage template else execute login template.
-	switch data.(type) {
-	case error:
-		data := setLoginDetails(false, false, "", "/login")
-
-		if err := loginTmpl.Execute(w, data); err != nil {
-			log.Println(err)
-		}
-
-	default:
-		if err := homepageTmpl.Execute(w, data); err != nil {
-			log.Println(err)
-		}
-	}
-}
 
 func HomePageLoginPost(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 	if err := r.ParseForm(); err != nil {
@@ -90,22 +32,6 @@ func HomePageLoginPost(w http.ResponseWriter, r *http.Request, _ httprouter.Para
 
 	w.WriteHeader(http.StatusOK)
 	sendResponse(w, token)
-}
-
-func validatUser(w http.ResponseWriter, r *http.Request) interface{} {
-	cookie := model.CookieDetail{CookieName: values.UserCookieName, Collection: values.UsersCollectionName}
-	if err := cookie.CheckCookie(r); err != nil {
-		return err
-	}
-
-	data := struct {
-		Email, UUID, Name string
-	}{
-		cookie.Email, cookie.Data.UUID,
-		values.MapEmailToName[cookie.Email],
-	}
-
-	return data
 }
 
 func setLoginDetails(errors, isAdmin bool, errorDetail, link string) struct {
