--- backend/views/admin/admin.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <meta name="description" content="admin portal">
-
-    <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet">
-    <link href="https://cdn.jsdelivr.net/npm/@mdi/font@4.x/css/materialdesignicons.min.css" rel="stylesheet">
-    <link href="https://cdn.jsdelivr.net/npm/vuetify@2.x/dist/vuetify.min.css" rel="stylesheet">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no, minimal-ui">
-
-    <title>Home</title>
-</head>
-
-<body>
-    <div id="app">
-        <v-app>
-            <v-content>
-                <v-container fluid>
-                    <div align="center">
-                        <tabs></tabs>
-                    </div>
-                </v-container>
-            </v-content>
-        </v-app>
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/vue@2.x/dist/vue.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vuetify@2.x/dist/vuetify.js"></script>
-    <script>
-        Vue.component('tabs', {
-            template: `(% template "tabs" .%)`,
-            (% template "tabsData". %)
-        })
-
-        new Vue({
-            el: '#app',
-            data: {
-
-            },
-            vuetify: new Vuetify(),
-        })
-    </script>
-</body>
-
-</html>
