--- backend/views/admin/components/block.vue
@@ -1,5 +1,0 @@
-(%define "blockUser"%)
-<template>
-  <v-card-text>Page Work In Progress</v-card-text>
-</template>
-(%end%)
