--- backend/views/admin/components/messagescan.vue
@@ -1,5 +1,0 @@
-(%define "messageScanner"%)
-<template>
-  <v-card-text>Page Work In Progress</v-card-text>
-</template>
-(%end%)
