--- backend/views/loginpage/login.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <meta name="description" content="login">
-
-    <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet">
-    <link href="https://cdn.jsdelivr.net/npm/@mdi/font@4.x/css/materialdesignicons.min.css" rel="stylesheet">
-    <link href="https://cdn.jsdelivr.net/npm/vuetify@2.x/dist/vuetify.min.css" rel="stylesheet">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no, minimal-ui">
-
-    <title>Home</title>
-</head>
-
-<body>
-    <div id="app">
-        <v-app>
-            <v-content>
-                <v-container fill-height>
-                    <v-spacer></v-spacer>
-                    <div align="center">
-                        <form action="(%.Login%)" method="POST">
-                            <v-row justify="center" align="center">
-                                <v-col cols="12">
-                                    <v-img src="/assets/unilag.svg" contain height="300" width="300"></v-img>
-                                </v-col>
-                                <v-col cols="12">
-                                    (%if .Admin%)
-                                    <span>Welcome Admin</span>
-                                    (%else%)
-                                    <span>Welcome to the University of Lagos chatting platform</span>
-                                    (%end%)
-                                </v-col>
-                                <v-col cols="12" sm="6">
-                                    <v-text-field v-model="email" width="100" type="text" full-width="false"
-                                        :rules="emailRules" name="email" rounded filled required
-                                        placeholder="Email Address">
-                                    </v-text-field>
-
-                                    <v-text-field type="password" rounded filled required name="password"
-                                        placeholder="Password">
-                                    </v-text-field>
-                                    (%if .SigninError%)
-                                    <v-alert dense text type="error">(%.ErrorDetail%)</v-alert>
-                                    (%end%)
-                                </v-col>
-                                <v-col cols="12">
-                                    <v-btn type="Submit">Log in</v-button>
-                                </v-col>
-                            </v-row>
-                        </form>
-                    </div>
-                    <v-spacer></v-spacer>
-                </v-container>
-            </v-content>
-        </v-app>
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/vue@2.x/dist/vue.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vuetify@2.x/dist/vuetify.js"></script>
-    <script>
-        new Vue({
-            el: '#app',
-            data: {
-                email: '',
-                emailRules: [
-                    v => !!v || 'email address is required for sign in',
-                    v => /.+@.+\..+/.test(v) || 'email address must be valid',
-                ]
-            },
-            vuetify: new Vuetify(),
-        })
-    </script>
-</body>
-
-</html>
